Protect a shared variable (chatId) from being incorrectly modified

+++ EmojiTelegramBot/Jobs/Webp2Gif.cs
@@ -3,24 +3,30 @@ using System.Threading.Tasks;
 
 namespace EmojiTelegramBot.Jobs
 {
-    /// <inheritdoc cref="IJob"/>
-    public class Webp2Gif : IJob
-    {
-        private string[] args;
+	/// <inheritdoc cref="IJob"/>
+	public class Webp2Gif : IJob
+	{
+		private string[] args;
 
-        public Webp2Gif(string[] args)
-        {
-            this.args = args;
-        }
+		public Webp2Gif(string[] args)
+		{
+			this.args = args;
+		}
 
-        public async Task<string> DoJobAsync()
-        {
-            string pathToGif = Path.ChangeExtension(args[0], ".gif");
-            await Task.Run(() =>
-            {
-                File.Move(args[0], pathToGif);
-            });
-            return pathToGif;
-        }
-    }
+		public async Task<JobResult> DoJobAsync()
+		{
+			string pathToGif = Path.ChangeExtension(args[0], ".gif");
+			var result = new JobResult(pathToGif, long.Parse(args[1]));
+
+			await Task.Run(() =>
+			{
+				if (!File.Exists(pathToGif))
+				{
+					File.Move(args[0], pathToGif);
+				}
+			});
+
+			return result;
+		}
+	}
 }
